Move ShowInBrowserService to IssueViz
